refactor: 예약 컨트롤러를 관리자와 사용자로 분리 및 엔드포인트 변경
